* commit 'd05f2aa1848ef4febcd25265d07ed92963c9e63f': (28 commits)
  >Parametrización challenge
  >Adaptacion queries nuevos índices
  >Correccion script
  >Cambios
  Arreglo consulta operaciones devolución y ventas
  > TIPO_GRUPO_DATO Asociado a las ventas que se ven desde la web
  Arreglo consulta de calcular importe comision venta o devolucion
  Corrección tests TipoPoliticaEnum
  Cálculo del importe comisión venta/devolución
  Totalización de operaciones de venta
  DML
  >Añadidos nuevos algoritmos/ajustes sin asignación
  Corrección test
  Correcciones localización pago y localización en cálculos por venta
  Cálculo del importe comisión vendedores (Por Venta)
  Eliminación del tipo de dato VENTA_INDIVIDUAL_LOCALIZACION (quedó en desuso) Creación del tipo de dato IMPORTE_COMISION_VENDEDORES_POR_VENTA
  Arreglo totalizar ventas empleados por venta
  Obtención de ventas tienda empleados por venta
  Totalización de presencias de personas por venta
  >Correccion desplazamiento base
  ...

+++ APPWEB-INCOMEMotorCalculoJava-model/src/main/java/com/inditex/rrhh/icmclcwb/model/primary/tarea/repository/TareaTipoHoraRepositoryCustomImpl.java
@@ -32,7 +32,7 @@ public class TareaTipoHoraRepositoryCustomImpl extends JdbcBatchPrimaryRepositor
         pstmt.setObject(3, entity.getIncluidoVenta());
         pstmt.setObject(4, entity.getIncluidoChallenge());
         pstmt.setObject(5, entity.getIncluidoEcommerce());
-        pstmt.setLong(6, entity.getIdTipoHora());
+        pstmt.setInt(6, entity.getIdTipoHora());
         pstmt.setLong(7, entity.getTarea().getId());
     }
